+++ valkey_glide_commands_common.h
@@ -222,6 +222,19 @@ int execute_client_unpause_command(zval*             object,
                                    int               argc,
                                    zval*             return_value,
                                    zend_class_entry* ce);
+int execute_latency_history_command(zval*             object,
+                                    int               argc,
+                                    zval*             return_value,
+                                    zend_class_entry* ce);
+int execute_latency_latest_command(zval*             object,
+                                   int               argc,
+                                   zval*             return_value,
+                                   zend_class_entry* ce);
+int execute_latency_reset_command(zval* object, int argc, zval* return_value, zend_class_entry* ce);
+int execute_latency_reset_with_route_command(zval*             object,
+                                             int               argc,
+                                             zval*             return_value,
+                                             zend_class_entry* ce);
 int execute_memory_doctor_command(zval* object, int argc, zval* return_value, zend_class_entry* ce);
 int execute_memory_malloc_stats_command(zval*             object,
                                         int               argc,
@@ -516,6 +529,19 @@ int execute_unlink_command(zval* object, int argc, zval* return_value, zend_clas
 
 #define REPLICAOF_METHOD_IMPL(class_name) \
     STANDARD_METHOD_IMPL(class_name, replicaof, execute_replicaof_command)
+
+#define LATENCY_HISTORY_METHOD_IMPL(class_name) \
+    STANDARD_METHOD_IMPL(class_name, latencyHistory, execute_latency_history_command)
+
+#define LATENCY_LATEST_METHOD_IMPL(class_name) \
+    STANDARD_METHOD_IMPL(class_name, latencyLatest, execute_latency_latest_command)
+
+#define LATENCY_RESET_METHOD_IMPL(class_name) \
+    STANDARD_METHOD_IMPL(class_name, latencyReset, execute_latency_reset_command)
+
+#define LATENCY_RESET_WITH_ROUTE_METHOD_IMPL(class_name) \
+    STANDARD_METHOD_IMPL(                                \
+        class_name, latencyResetWithRoute, execute_latency_reset_with_route_command)
 
 #define MEMORY_DOCTOR_METHOD_IMPL(class_name) \
     STANDARD_METHOD_IMPL(class_name, memoryDoctor, execute_memory_doctor_command)